+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add response viewer functionality and enhance UI elements. (`5da28a40`)
 - Refactor LinuxAppSubmenuItem to improve nested flyout padding. (`f4d9ceb8`)
 
 ## 2.8.4 - 2026-07-27